TST: Move effective permeability to `test_utils`
